Automatically set driver version from artifact version
The version of the driver was hard coded in AthenaDatabaseMetaData which causes the tests to fail when bumping the version during the release process. To fix this we create a properties file, and tell Maven to write the project version to this file. Then we load and parse the version number from this properties file.

There are also another properties file `META-INF/maven/io.burt/athena-jdbc/pom.properties` added automatically to the jar file, but this file cannot be used in the unit tests.